feat(utils): 添加 useRouter 工具函数
