+++ src/App.vue
@@ -1,38 +1,18 @@
 <template>
   <div id="app">
-    <div>{{ state.value }}</div>
-    <button @click="dec"> - </button>
-    <button @click="inc"> + </button>
+    <router-view></router-view>
   </div>
 </template>
 
 <script>
-import { reactive, watchEffect } from '@vue/composition-api';
+import { provideRouter } from 'utils/use-router';
 
 export default {
   name: 'App',
   components: {
   },
-  setup() {
-    const state = reactive({ value: 0 });
-
-    const inc = () => {
-      state.value += 1;
-    };
-
-    const dec = () => {
-      state.value -= 1;
-    };
-
-    watchEffect(() => {
-      console.log(state.value);
-    });
-
-    return {
-      state,
-      inc,
-      dec,
-    };
+  setup(props, context) {
+    provideRouter(context.root.$router);
   },
 };
 </script>
